Add ability to disable HA for non-Ceilometer queues
Add a flag enable_rpc_ha which enables queue mirroring for all queues.
Setting this flag to false disables mirroring for RPC queues. Note
that it does not affect mirroring for Ceilometer queues, which are
always mirrored.

Since the feature is experimental, the flag is set to true by default
to preserve current behaviour.

Text for Operations Guide

It is possible to significantly reduce load which OpenStack puts on
RabbitMQ by disabling queue mirroring for RPC queues. This could be
done by executing the following command on one of the controllers:

    crm_resource --resource p_rabbitmq-server --set-parameter \
        enable_rpc_ha --parameter-value false

This will not affect Ceilometer queues, as they are always mirrored.

Note that it is not yet well tested how this configuration affects
failover when some cluster nodes go down. Hence it is experimental,
use at your own risk!

DocImpact:  ops-guide

Implements: blueprint rabbitmq-disable-mirroring-for-rpc
Change-Id: I80ae231ca64e2a903b0968d36ba0e85ca9cc9891