Add ability to disable HA for non-Ceilometer queues
Add a flag enable_rpc_ha which enables queue mirroring for all queues.
Setting this flag to false disables mirroring for RPC queues. Note
that it does not affect mirroring for Ceilometer queues, which are
always mirrored.

Since the feature is experimental, the flag is set to true by default
to preserve current behaviour.

The change is implemented in several steps:
 * the upstream script changed so that it allows to extend the
   list of parameters and uses a policy file to define RabbitMQ
   policies.
 * we add our own version of OCF script which wraps around
   the upstream one. It defines a new enable_rpc_ha parameter and
   passes its value to the upstream script.
 * we add our policy file, where we use the enable_rpc_ha parameter
   to decide which policies we should set.

So we will have two OCF scripts for RabbitMQ in our deployment:
 * rabbitmq-server-upstream - the upstream version
 * rabbitmq-server - our extention, which will be used in the
   environment

The upstream version of the script will be pushed to the upstream
along with empty policy file, so that other users can define their
own policies or extend the script if needed. Here is the pull
request: https://github.com/rabbitmq/rabbitmq-server/pull/478

Text for Operations Guide

It is possible to significantly reduce load which OpenStack puts on
RabbitMQ by disabling queue mirroring for RPC queues. This could be
done by executing the following command on one of the controllers:

    crm_resource --resource p_rabbitmq-server --set-parameter \
        enable_rpc_ha --parameter-value false

This will not affect Ceilometer queues, as they are always mirrored.

Note that it is not yet well tested how this configuration affects
failover when some cluster nodes go down. Hence it is experimental,
use at your own risk!

DocImpact:  ops-guide

Implements: blueprint rabbitmq-disable-mirroring-for-rpc
Change-Id: I80ae231ca64e2a903b0968d36ba0e85ca9cc9891