Fix fence_ssh python script
Import module atexit explicitly. Since a lot of packages were updated
during migration to CentOS7 this script no more works without that.
Probably it was imported implicitly, which is wrong anyway.

Fix command line options changes according to updated fencing.py

Remove test if needed shutdown slave because in any case all
that we do is just restart the slave. Always.

Blueprint: master-on-centos7

Is NOT compatible with CentOS6 master node

Change-Id: I0437a71ac59a46eac58714bb28d78882987e5938