Change protocol and hostname for novncproxy
In order to provide VM console access via HTTPS, we need to provide
the correct link to the proxy via the api.
Also, the cloud administrator will need to properly configure the DNS
for the clients to ensure they will be able to access the HTTPS
endpoint as the DNS name must match the certificate name for the Nova
VNC Proxy service.

Change-Id: I99be041b547284edc57b76179f8acc2566662307
Closes-bug: #1511422