Bump Glance ref on a new tag
Update Glance ref on a new tag, because was merged upstream change [1],
which closes bug (#1499754).

Currently during deployment Glance (with Swift as backend) we have
reference to every glance image in database with credentials, like:
swift+http://services%glance:EF3tluYM@192.168.0.2:5000/v2.0/glance/049cd701

This is not correct, because if someone will change glance credentials in conf
file, this reference will be invalid.

[1] Iaf0c1c6f2188c84267e41d2d4f9bda0dbf3dccad

Closes-bug: #1499754

Change-Id: I7e80f43a8e7664eeb469a7627022c37498be712c