Added cross-dependencies to pre-deployment tasks
We have a set of tasks which have task from master node
in "requires" list. Actually these dependencies are
cross-node ones but it works in current granular-deploy
implementation only because all pre- and post- deployment
tasks are executed one-by-one for whole cluster. If we try
to use task-based deployment we need to add these
dependencies as cross-node ones.

Implements blueprint: task-based-deployment-astute

Change-Id: I57e845f4834d51e1a8437a37c576c6dde7a82d8c