Configure keymgr auth url if identity_uri is used
Cinder requires a keystone URI for performing quota operations.
This behiavor is present in Liberty release, but will be fixed
in Mitaka release to re-use the identity_uri parameter.

This change adds the parameter keymgr_encryption_auth_url to
identity_uri + /v3 if identity_uri is set when calling cinder:api.

Updates openstack-cinder ref to mos-8.0 to pull in
https://review.fuel-infra.org/#/c/14476/

This can be reverted after LP#1516085 is in use for Mitaka cycle.

Closes-Bug: #1521640
Change-Id: Ide9b35761b25be155d348bf8c8704efc23634482