Install heat-docker package only after heat-engine
Heat-engine package is heat-docker dependency,
so if installation of heat-docker package is performed before
heat-engine installation it leads to autostart of heat-engine package
(in that moment override file isn't exist yet, because it will be
triggered directly before evaluation of heat-engine package resource).
As a result, the heat-engine service is run without any configuration.

To fix that we need to sеt heat-engine package installation strictly
before heat-docker package.

Change-Id: I5420e64e3ab6b2ca0305f5f41eb722e3ead42b25
Closes-bug: #1520610