Fix detached keystone node epoch mismatch
Pacemaker maintains an internal database, which is used for
configuration storage. Each update of this database increases a counter,
called "epoch", which should have the same value cluster-wide.

If an update operation comes to a previously detached node, a conflict
will occur. Pacemaker does not allow updating this database on a node,
which epoch value is lower than the epoch value of a cluster leader.

We should wait the epoch counter to come into sync by perodically
retrying the update command.

Closes-Bug: 1494314
Change-Id: I1f242bcd90264ec45da2aaa6bc030f244511761b