Install Murano artifact repository for Glance
This package adds a definition of murano-specific
artifact type, which acts as an extension of Glance
v3 API, enabling it to store, process and query
murano-specific artifacts. Since Glance uses stevedore
to discover its artifact types as plugins, this type
has to be installed into the same pythonic environment
with glance-api. When the v3 is disabled if glance-api
configuration, this package is not used.

This package can be used by another applications
(not only by Murano) that's why it should be always
installed (and shouldn't depend on Murano enablement).

Depends on:
https://review.fuel-infra.org/#/c/14675/
https://review.openstack.org/#/c/259483

DocImpact

Change-Id: I48cf5f00d2c9fe9324bf3d6e91ce2adb9f9ea20e
Closes-bug: #1525289
(cherry picked from commit 7f7bad1a14e95f19e0f1c2ce67467c4d0491ffa8)