Change apache2 CGI module for RadosGW
In apache-2.4.7 fastcgi module is not support chunked transfer.
It breaks uploads of objects to RadosGW.
Fix available since apache-2.4.10 (vivid) and upgrade apache in MOS is not
possible, so this workaround proposed.
RadosGW runs with native CGI server on 127.0.0.1:9000 and apache uses
mod_proxy_fcgi to proxy user requests to rados.

Closes-bug: #1521623
Closes-bug: #1505273

Change-Id: Ia3959c4009c6c673074a9b7536b1cc87834c301d