Update connection string generation
This change introduces a generic function for constructing database
connection strings for use with all of the OpenStack services.  As part
of this change, we are moving the database connection string generation
to the osnailyfacter modular tasks and updating the openstack classes to
consume a connection string rather than the seperate host, user,
password, database name and database type.

Additionally as part of this change, we are updating the mysql
connection strings to only include read_time=60 if debian based packages
are used. This is due to the fact that python-pymysql, which is used by
Ubuntu packages, does not support the read_timeout option as part of the
connection string.

Change-Id: I45c51a05ba7cf9648b0202cc3c2216d185bd7eda
Partial-Bug: #1526938