Reduce database idle timeout to 300 seconds
This change standardizes the way we define the database idle timeout,
max pool size, max overflow and max retry settings. With this change we
are now setting the idle timeout to 300 seconds from 3600 seconds to
help reduce failures when failovers occur. With the smaller timeout,
there should be less idle connections that would be affected when a
database server goes away.

With this change, we are also allowing for the 4 database options to be
overriden using the hiera data for the respective services.  For
example, you could tune keystone with a higher (or lower) timeout then
the other serivces by defining database_idle_timeout in the keystone
hiera hash.  The following keys can be used in the respective service
hashs to override the settings:

  database_max_pool_size
  database_max_overflow
  database_max_retires
  database_idle_timeout

Change-Id: I1506e0a2e01dde3556777011a72967b698955d99
Closes-Bug: #1525214