Debug and Notice messages during Fuel master start
Instead of running 'dockerctl build all' command in the background
and long wait for a debug output we've split the command to build
each container one by one, but keep the original order. Doing this,
we have much more information what is going on.

Also this commit adds some important message to the log which used
as an anchor for VirtualBox scripts.

Change-Id: I77a1d3a32633390c1fe37bbeb81da5d0996972df