Fix the proc_kill ocf helper func
W/o this fix, the command "proc_kill pid name count" will misplace
the passed count of tries as the signal name for the pkill, which
is wrong and might be unpredictably harmful for the proces group
containing the given pid.

The fix is to correctly specify params for the proc_kill calls.
Add bats tests for the proc_kill as well.

Closes-bug: #1528889

Change-Id: I992480f4e0d3380215d3f8bd910553070334c343
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>