Configure keystone cache memcache servers
This change properly configures the cache/memcache_servers configuration
for keystone. Due to an upstream keystone module bug, this setting is
not properly configured and there is no support for configuring it
within the keystone module at this time. Without this setting, the
keystone cache is enabled without properly configuring the memcache
servers which results in errors and excessive backend communications.

Change-Id: I6ca90dfb32c824aeff8022ba70d7b988edae6486
Partial-Bug: #1530178