Improve rabbitmq OCF script diagnostics
Currently time-out when running 'rabbitmqctl list_channels' is treated
as a sign that current node is unhealthy. But it could not be the case,
as the hanging channel could be actually on some other node. Given that
currently we have more than one bug related to 'list_channels', it makes
sense to improve diagnostics here.

This patch doesn't change any behaviour, only improves logging after
time-out happens. If time-outs continue to occur (even with latest
rabbitmq versions or with backported fixes), we could switch to this
improved list_channels and kill rabbitmq only if stuck channels are
located on current node. But I hope that all related rabbitmq bugs were
already closed.

It doesn't make sense to push this patch upstream - it's just a hack for
older versions of rabbitmq that we are using now. Real improvements will
go into 'rabbitmqctl list_channels' itself, somewhere in post-3.6.0. But
we will not be able to backport them - there was too many changes about
'list_*' commands.

Change-Id: I4746d3a4e85dc2a51af581034ae09a1cf0eefce2
Partial-Bug: #1515223
Partial-Bug: #1513511