Skip docker resource install with UCA packages
This change updates the heat modular task to only attempt to install the
heat docker packages when using a package set that provides a standalone
package. Currently only the RedHat and Debian based package sets provide
a seperate package. The UCA packages do not provide the heat docker
templates as a separate package.

Change-Id: If66f928f8afdebd3eedad1da9ac9e5c1e23c2101
Related-Blueprint: deploy-with-uca-packages