Fix horizon caching backend for UCA
This change updates the horizon caching backend to be configurable and
changes the default where os_package_default is either not present or is
set to debian. The MOS package provides an alternate memcache backend
that is not available from other groups. With this change, we can now
specify a cache_backend as part of the horizon configuration hash which
will be used instead of the defaults specified in code.

Change-Id: If1050db4bb54bc9f3ae355ffeb2ae7fbeb3cfc44
Related-Blueprint: deploy-with-uca-packages