Limit max service workers by RAM
Intoduce the worker RAM ratio to define the total RAM every single
worker of all service types may consume. More services share the
same node, more RAM ratio should be given to the workers.
The default value assumes there are 20 different types of workers
limited by 100Mb each (or 10 types by 200Mb and so on).

(tl;dr)
Without this fix, we use hardcoded thresholds for service
workers (process/eventlet based) of 16 processes. And for rabbit -
of 100 threads. While the actual values for workers are evaluated
as only CPUs based, with the hardcoded thresholds applied.

This is a problem as the limiting global tresholds must depend on the
physical RAM as well. We don't want 16 workers for an 8CPU, 1G RAM node
to be the same as for an 8CPU & 128G RAM one. Note, that this problem
has nothing to the custom service_workers configurable per each service
in hiera. The global threshold applies only when those are not defined
as well.

The solution is to introduce the workers_ratio (configurable in hiera).
This represents the average workers-to-RAM ratio for a multi process
workers of each known type. The default value is 2000, which is like
10 generic workers of different services to consume 200M each,
or the same as 20 workers X 100M, if there are 20 distinct services.

The workers_ratio serves as an argument for a new global treshold
function limiting the maximum amount of workers for the each distinct
service type. So, there will be no more hardcoded limits of 16 / 100.
Instead, those will be evaluated as a function F(RAM, workers_ratio).
Examples:
F(1G, 2000)=1 global workers threshold for processes;
F(2G, 2000)=2;
F(8G, 2000)=5;
F(64G, 2000)=F(128G, 2000)=30

And the service_workers will be evaluated as:
$threads = min($workers_max,max(12*$physicalprocessorcount,30))
$processess = min(max($::processorcount, 2), $workers_max)
Instead as it was before:
$threads = min(100,max(12*$physicalprocessorcount,30))
$processess = min(max($::processorcount, 2), 16)

(Note that the effective threads value will be doubled for the rabbitmq
thread workers. We don't change other thread workers, like apache2, as they
are implemented to depend on the available RAM already)

Also, add missing noop tests for service workers, address todo for the
rabbitmq in the openstack::nova::controller class (no more needed as
was removed).

Closes-bug: #1529096

Change-Id: Ib90838cff05419d824d064e93b13b627b5a07fea
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>
(cherry picked from commit aa999f7f5b3e75f5f5af1f17d0586edb88c9dc8f)
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>