Update puppet-keystone and puppet-heat modules to currnet Mitaka version
Puppet-keystone and puppet-heat modules don't contain any custom changes.
Keystone module is up to date with upstream version of 11 Jan.
Heat module is up to date with upstream version of 30 Dec.

Update of this modules can't be separated because of changes in
keystone_user provider and heat user/domain creation.

Additional changes:

  * Remove public_endpoint definition:
    Mitaka manifests treat false as a string False, rather
    than unsetting the value. This parameter is no longer
    used, so it should be simply removed.

Change-Id: I575d2811160796b07f4aaeece466f582bc2d788a