Fix detached keystone node epoch mismatch
Pacemaker maintains an internal database, which is used for
configuration storage. Each update of this database increases a counter,
called "epoch", which should have the same value cluster-wide.

If an update operation comes to a previously detached node, a conflict
will occur. Pacemaker does not allow updating this database on a node,
which epoch value is lower than the epoch value of a cluster leader.

We should wait the epoch counter to come into sync by perodically
retrying the update command.

Change-Id: I6334ea7cbc569ab2c14a44af8da44158f6f629c2
Closes-Bug: 1494314