Change apache api proxy  config to disallow keepalives for OSTF
Sahara API does not support keepalive connections correctly, so it is
necessary to disable this feature. It is already disabled for HAProxy,
but in case with OSTF tests there is Apache working as another proxy
between admin and public network, so we should disable keepalives for
it too. In other case there is cases when client tries to reuse
connection for HTTPS request and fails to do it.

Closes-Bug: #1533218
Change-Id: I9cb4f69e49c8aa71b861e2cc1ee285f6d23dc4c1
Co-Authored-By: Dmitry Bilunov <dbilunov@mirantis.com>