Wrap ntpdate in pre-deployment to timeout
Mcollective send TERM signal on retry, which can lead to problems
with non-terminated application that hangs and consequent blocking
second instance of that application which will lead to non-zero
exit and deployment blocking. Instead this we should send KILL manually
on timeout to be sure that we stop application and next instance won't
be blocked.

Change-Id: If2967106ef7267a454ca012a444eeda9b7c05793
Closes-Bug: #1533082