Specify backend name for all Cinder backends
Currently for configuration cinder backends we are using volume classes,
and all of them have volume_backend_name 'DEFAULT'. It will lead to
problems, when we have multibackend for Cinder.

To avoid this problem, this patch switches using volumes to backend
defines, which allow to pass volume_backend_name.

Because Puppet 3.x doesn't support to handle array, in this patch
was added custom define, which call upstream cinder classes, that
create cinder types.

Change-Id: I24023b01e1472f82197aa5c4807cb2728c67d742
Closes-bug: #1536643