Fix ml2_conf.ini to plugin.ini link
The Kilo packages expect /etc/neutron/plugin.ini to be linked to the proper
config file bfore the ovs agent service can start. This was removed in more recent version of the puppet-neutron modules. We need to create this link untill the neturon package can be updated to point to the correct startup config files

Change-Id: Iea413e3f37beb628b44655e289c7b3c004f0e38f
Related-bug: #1537941