avoid to premature restart nova-compute by ceilometer
There is ceilometer-compute compute task that configures ceilometer
and restarts nova-compute in order to apply new settings. It is totally
correct behaviour for compute node. But for compute-vmware it restarts
unconfigured nova-compute. As a result nova-compute starts
with the wrong parameter host (nodename instead vcenter-servicename)
and registers in the nova database with that paramter. Then
top-role-compute-vmware sets right value for host and restarts
nova-compute that also registers in the database.

Therefore we got the record in the database for nova-compute and host a
equal nodename. When ostf test checks that all process are alive it
found this record and decides that it's dead. And test failed.

This commit allows to avoid premature restarts nova-compute before it
will configured and solves this problem.

Change-Id: Ica3136423303c4ed9f36a194080d3489a793f980
Closes-bug: #1529840