Add support for auto libvirt type detection
This change adds a fact called virtualization_support which checks for
the virtualization flags in the cpu info (vmx|svm) and sets the fact to
'true' or 'false' depending on the cpu flags existance.  libvirt_type is
set to 'auto' in hiera, we will use the fact to determine if we should
use kvm or qemu. This change still allows a user to specify a specific
libvirt_type to bypass this auto detection.

Change-Id: I017965929e95482b52b1dd457f211b18c3f01721
Related-Bug: #1536836