Change protocol and hostname for novncproxy
Noop test, actual fixtures for this changes and other VMware specific
changes we provide asap.

In order to provide VM console access via HTTPS, we need to provide
the correct link to the proxy via the api.
Also, the cloud administrator will need to properly configure the DNS
for the clients to ensure they will be able to access the HTTPS
endpoint as the DNS name must match the certificate name for the Nova
VNC Proxy service.

Change-Id: Ic8da073bb54241c4533747687d106f5f58ffdb4b
Closes-bug: #1511422