[PATCH] Update Glance Swift backend configuration Currently, when Glance uses Swift as backend, all references to images saved in database with credentials, like: swift+http://services%3Aglance:EF3tluYM@192.168.0.2:5000/v2.0/glance/049cd701-77
The problem is that when we want to change Glance credentials we can't use
old images anymore, because new credentials and creds in database are
different. For resolving this issue was implemented patch in Glance [1]
and couple of patches in Glance_Store [2,3]. The main goal is using
separate section (file with such section) for providing credentials
for access to Swift and do not place reference with credentials to
database.

All changes above are covered in docs for configuring Swift backend [4].

Also glance_cache_config now is not using for configuring Swift backend,
hence all related parameters for this config removed.

[1] https://review.openstack.org/#/c/98722
[2] https://review.openstack.org/#/c/193422/
[3] https://review.openstack.org/#/c/108036/
[4] https://review.openstack.org/#/c/242583/

Change-Id: Iaf0c1c6f2188c84267e41d2d4f9bda0dbf3dccad
Closes-bug: #1499754