Allow ceph-disk activate-all to return errors
`ceph-disk activate-all` is used in conjuction with `udevadm trigger`
to prevent allready prepapred/provisioned disks from being re-formatted.

The error expressed in bug: 1541946 is that activate-all has errors
because there is no filesystem on the devices. This is accurate on nodes
where the OSD was not previously sucessful. To combat this we should
allow for the command to return errors

Change-Id: I5cf29fcba7b0443832489c90518a9197c065ac8c
Closes-bug: 1541946