Fix openstack-cinder idempotency
This change updates the openstack-cinder task to be completely
idempotent on the second run.  Previously we were improperly includeing
the ubuntu service tweak for cinder-backup for non ceph configurations
which was resulting in the extra execution of the policy-rc.d execs.
This change wraps the cinder-backup tweak in a check for ceph to prevent
the extra items being changed on the second run.

Change-Id: Ifb9c6082fac93f65d50cf949966588187c2441fb
Related-Blueprint: granular-task-idempotency