Make heat task idempotent
This change updates the heat task to be idempotent. In order to do this,
we have changed the wait_for_heat_config exec to be refresh only and is
now only triggered if configuration changes are done. Additionally this
change cleans up the pacemaker integration for heat and moves it into
the cluster module. With the addition to the cluster module, unit tests
have been added to the new cluster::heat_engine class.

The resources that were always being executed prior to this change:
 - Cs_resource[p_heat-engine]
 - Exec[wait_for_heat_config]

Change-Id: Ie0f7137206f2733259bff7b0c2c86fb1634add76
Related-Blueprint: granular-task-idempotency