Refactor openstack-controller task
This change refactors the openstack-controller task to remove the
openstack::controller module and pull the nova configuration straight
into the task. As part of this refactoring we are also addressing the
idempotent issues around the use of tweaks::ubuntu_service_override for
packages and services we are not including.

Additionall we are also addresssing the fact that the
default_floating_pool was not being properly configured durring this
task which could result in nova misconfiguration if this task was
executed after a deployment.

Change-Id: I2acd443f60f9bebdc828d59aecd11cf453fc8c67
Related-Blueprint: granular-task-idempotency