Fix `dockerctl start all' not working with systemd
There is a missing local variable assignment for "$container" in
./files/fuel-docker-utils/functions.sh. Instead an environment
variable leaks from /usr/bin/dockerctl into functions.sh's
start_container and it tries to start unit "docker-all" which
obviously does not exist.

This patch assigns the first argument of start_container to a
local variable "$container" which fixes this regression.

This bug cannot be reproduced on stable/8.0. It can be triggered
least in fuel-9.0-liberty-195.

Change-Id: I1a9c7e965b37df90f81f14ddad1513d2914c14ad
Closes-Bug: #1543110