Link neutron syncdb with service run.
To have an ability for API calling, we must have neutron fully
configured, so we must configure all resources, sync DB, then
run neutron.

We could move syncdb task to server-config manifest, but it is
not good solution, cause it will be tricky to link it proper way. So,
just run it always before neutron service.

Change-Id: I528ee55f1766d59bd2320b59bd4e181b4df350ed
Related-Blueprint: granular-task-idempotency