Disable HA for RPC queues by default
RPC wihout HA was tested on a big scale and we found that it greatly
reduces load OpenStack puts on RabbitMQ. Hence it is valuable to
disable it by default.

DocImpact

It should be noted in the release notes that starting from 8.0 queue
mirroring is disabled by default for RPC queues, but it is still
enabled for Ceilometer ones.

Also, the change should be reflected in our reference architecture
guide. Specifically, there is a sentence here starting with
"RabbitMQ provides active/active high availability ...", which will
be incorrect after the given change is merged.

Users are still provided with means to enable mirroring. Details
could be found in description of that commit:
https://review.openstack.org/#/c/249180/

Closes-Bug: #1550303
Change-Id: Iffa4173c2e6bb54e411defc7bdc44254669be5fd