Use systemd journal for stdout and stderr
The thing is that when Fuel services were
deployed inside docker containers it was inconvenient
to use several journals and we used syslog for
stdout and stderr channels. At the moment
all services are deployed on the Fuel master
node host and there is no need to use syslog.
It is better to switch to systemd journal which
is used by default.

Closes-Bug: #1544194
Change-Id: I419570ffdc002251d25298592f2ac76e00972310