[DNF] Switch from jbussdieker/monit to sbitio/monit
The current module doesn't specify the license, making it impossible
to include it in Debian.

Due to problematic naming scheme of mirrored modules on Fuel infra,
the switch will require forced push to mirror in order to actually
switch to new upstream. This commit is a proof that the main change[1]
passes CI tests.

[1] I259a83e56f488a2a304f2f2410fc5b16719a969f

Change-Id: Ic27a6c8f84984544ade4a134648028a44403aad0
Related-Bug: 1544428