Disable globals.yaml when globals task ran
Variables names in globals.yaml and astute.yaml are the same sometimes.
It leads to bad lookups when we run globals.pp for a second and subsequent
times. To avoid this previously we deleted original globals.yaml file
each time globals.pp was ran, which loads us to idempotensy loss even in
case of noop run. It was a bad approach, so to save previous behaviour
but avoid file deletion we change default globals source in hiera to
dynamically defined and redefine it in globals.pp manifest.

Change-Id: I35cc5441001856a8498be6c7ea515685e97ae90c
Related-Blueprint: granular-task-idempotency