Fix rspec tests for the 'l23network' module
This commit fixes the second non-critical part of the bug with
regular expression warnings and eliminates ugly messages during
the exception in the 'route_delete' function.

Also, it contains the fix for the 'package' resource with the
attribute 'ensure' set to 'purged' value when RSpec tests executed
on non-Ubuntu environments.

Change-Id: Ibfad0fb4142dcbd5de086183caa805db50e9c86e
Partial-Bug: #1539975