Fix access for detached database plugin
With the switch to the upstream mysql/galera modules we lost the
additional grants used by the detached database plugin to allow for
database creation and management from other hosts. This change brings
back the additional grants for access along with tests.

Additionally with this change we are adding an ensure_resource for the
/root/.my.cnf because when the detached database used the file does not
exist in the catalog to be switched to a symlink with the spaceship
operator.

Change-Id: I93c985db86ea1f1d27396ee0ef2920b9a9ee2802
Closes-Bug: #1547150