Ensure parent directory exists for auxiliary CentOS repo
In case of clean CentOS system we don't have directory structures
for auxiliary CentOS repos, so let's create it correctly.

For now it's not so simple to track directory tree structure,
since several classes can add some "custom" variables to path,
so for now the only case is to use "mkdir -p $some_path" solution.

Change-Id: I437e559abf6d71c355a434591127cd825a665207
Closes-bug: #1547504