Fix expected input to osnailyfacter::dnsmasq, ntp and ntp_available
The osnailyfacter::dnsmasq expects the external_dns parameter to be an
array. This change adds validation to the dns server list as well as
adding support for a single ip address provided as a string. If the list
of dns servers is not an array or a single ip address (as string), try
to convert input data to array from string.

Also do the same for ntp_available function and ntp::init class.

Change-Id: I6152b8d73a7b97b30d1d1fed9c8d1f055a9d933b
Closes-Bug: #1547654