[WIP] Fix database.pp idempotency for controllers
This change adjusts the ordering of the .my.cnf configuration for the
non-primary controllers as they already get a database with a root
password that isn't blank initially. So we need to try and put .my.cnf
after the mysql server starts and before the mysql root password class.

Change-Id: If3d497c281c1a402c68492afb1c638938d512804
Related-Blueprint: granular-task-idempotency