Start rabbitmqadmin on localhost for master node
Upstream puppetlabs-rabbitmq module should be updated to support
different bind_ip definition both for rabbitmq and rabbitmqadmin.
For now two variables can be used:
   * node_ip_address
   * management_bind_ip, for example:
      $rabbitmq_management_variables = {
        'listener' => "[
           {port, ${management_port}},
           {ip, \"${management_bind_ip}\"}
        ]"
      }

so there can be a situation when:
    $node_ip_address != $management_bind_ip
and as the result 'rabbitmqadmin' can fail

This patch set management_bind_ip to localhost, since it's check
by default in upstream module [1] we used. Switching to new
tag [2] will not help since for now check depends on
$node_ip_address [3] which can differ from $management_bind_ip

[1]. https://github.com/fuel-infra/puppetlabs-rabbitmq/blob/\
             5.3.1-mos-rc1/manifests/install/rabbitmqadmin.pp#L18
[2]. https://github.com/fuel-infra/puppetlabs-rabbitmq/blob/\
             5.3.1-mos-rc2/manifests/install/rabbitmqadmin.pp#L19
[3]. https://github.com/puppetlabs/puppetlabs-rabbitmq/pull/432

Related-bug: #1548755
Change-Id: Ia7311c70a84f01715144942616f7dcac14b1b9ce