Fix mysql upstart service
Upstart has not disabled starting process along with pacemaker

Puppetfile updates puppet-pacemaker to disable basic service not only on
start but on status also. This eliminates the problem when service is
started but name.override file is deleted manually. Thus we insure that
override file is always present on every puppet run.

Change-Id: I95bc68170dc4d72ae4914cb27dc304a96251c2a0
Closes-bug: #1550929
Signed-off-by: Sergii Golovatiuk <sgolovatiuk@mirantis.com>