Add SR-IOV availability check into task graph
We need to check if SR-IOV is properly enabled on the node before
we try to configure it. In order to do so and to fail early this
new ruby shell task is added. It checks SR-IOV availability before
running any Nova configuration tasks on compute nodes.

Change-Id: Ide5dfc4779864442b0765e1101d67240398b2f18
Implements: blueprint support-sriov