Timeout variable and finite number of checks were added to mysql_start
"If there is one cloned mysql instance not starting, the
whole pacemaker cluster gets stuck and does not emit any log. On the
host of the failed instance, I can see a mysql resource agent process
calling the sleep command."

To solve this problem the number of checks that server came up after
starting mysql was bounded in mysql_start method of MySQL RA OCF.
Also there was added the startup timeout variable.

Change-Id: I0789d5aedfa6d8e0592881b073e0ae4b8ac0d1fe
Closes-bug: #1449542