Fix race condition around initial sync
This change updates the wait for sync to only continue when galera is
reporting that the node has been Synced. Previously we were allowing it
to continue if the state was reported as Initialized as well. This leads
to task failures because mysql may not be ready for access when status
is reported as Initialized. Additionally, this change updates the
database task to be run one_by_one to reduce the race conditions with
the nodes joining the cluster at the same time.

Change-Id: I04956d5b875876b3b0d34a80dff0528bbe7ed76d
Related-Bug: #1552153