Set proper public address for swift endpoint for Ceph
Currently we are not handling situation when we have Ceph with RadosGW
enabled with enabled TLS/SSL. As a result we have wrong addresses
for Swift endpoints in Keystone. This commit fixes that.

Also this patch creates radosgw_keystone task for managing endpoint and
removes unused parameters from ceph-osd and radosgw modulars.

Change-Id: Id86d85a13fbec27265156a5b12fad0fd892c7cd9
Closes-bug: #1540133