Add ignore_db_dirs support to Galera
Previously we were setting ignore_db_dirs in the wsrep.cnf file for
mysql. This is incorrect as this setting should be in the my.cnf.
This change adds the ignore_db_dirs setting to the galera module
which manages the my.cnf for the fuel galera deployment.

Change-Id: I91399d8899b999622138f8408908ef0b2c5f1cf6
Closes-Bug: #1484552
Signed-off-by: Maksim Malchuk <mmalchuk@mirantis.com>
(cherry picked from commit f302cb1ee1d55a24036b620ff01d57ca2e30014c)