Fix noop and alarm hsitory ttl for oadh
The alarm history ttl has been removed from ceilometer and should now
be set as part of the aodh configuration. This change removes the
check from the ceilometer tests and implements it for aodh.

Also change I48ec5711a507115d8b7cc167ae881cac4637622f removed the
*_hash variables but Change I30a3832a2277fcf54396cb9b13da6050adf85edd
added in node_hash which is now no longer defined. Since the latter
change landed after the removal of the *_hash vars and we do not gate
on noop we ended up with broken noop tests.

Change-Id: I37a7354dfdc5a9592fa244bc87f8ab4f3f924bdf
Closes-Bug: #1554146
Closes-Bug: #1554168