Enable high precision timestamps for syslog
This change updates the logging manifests to enable the high prevision
timestamps for the fuel master as well as the deployed node. Previously
our custom rsyslog template had different logic around the
high_prevision_timestamps setting then what is in the upstream rsyslog
module. This change updates our invocations of the rsyslog::server and
rsyslog::client to enable the high precision timestamps.

Change-Id: I875fa2201fb0c66835c29bb5faadf8e03ae7a794
Closes-Bug: #1554114