Fix cinder type ensurabilty
This change adds additional logic to our cinder type management to
allow for the automatic cleanup of cinder types when they are no longer
configured. Previously we were only creating types and when we unlock
the ability to reconfigure an environment, we would leave the old cinder
types still configured. A user would have to manually cleanup the cinder
types via the 'cinder type-delete' command.  This change replaces the
usage of the cinder::type resource with the new cinder_type provider
that allows for ensure => 'absent'. The old cinder::type has also been
deprecated so this removes its usage.

With this change we are also moving the resource we were using to a
proper resource with tests under the osnailyfacter module.

Change-Id: Ib547c4c76b89cab10ccbcefc021f62b621a79c08
Related-Blueprint: granular-task-ensurability
Related-Bug: #1555360