Refactor out opensack::swift::* classes
We have tradtionally employed a combined storage and swift proxy role/task
but they are kept seperate in both our aged openstack::swift::* and in the
swift module its self.

In this review, we will refactor the single swift.pp task to separate proxy.pp
and storage.pp tasks. In these tasks, we are only altering what was previously
in swift.pp with the inclusion of the openstack::swift::* classes directly in
task. In order to minimise review burden, we will follow up with another
review to flatten out the openstack::swift::* calls

Change-Id: I1d5d6024a7491b0c58bbd65f958d2f61e797233c
Blueprint: fuel-remove-conflict-openstack