Patch thd_get_ha_data in mysql-server-wsrep-5.6
@@@@@@@@@@@@@@@@@
@ DO NOT MERGE! @
@@@@@@@@@@@@@@@@@

Function thd_get_ha_data is called without ensuring that the
corresponding slot in the thd structure actually exists.

This patch makes thd_get_ha_data return NULL if the current slot,
specified by handlerton structure (2nd argument) has not been populated
yet (it equals -1).

Do not merge this patch because it relies on a specific build of
mysql-server-wsrep-5.6 and a default value of gcc's -falign-functions.
To perform the actual check we need some extra space in the text
segment. This patch abuses the inter-function gap filled with x86-64
long nops (20 bytes) and the fact that high dword of rax is set to zero
by the callee to save an extra byte.

Change-Id: I180e9c4a17a007e73af3ce4c854a9d6259d59c5b
Partial-Bug: #1541338