Update keystone db task to wait for all databases
This change updates the keystone db task to wait for the database tasks
on all nodes to complete before getting run. I believe that this is
causing the galera cluster to become corrupt because the following
condition occurs:

- primary-db comes up
- secondary-db comes up and syncs from primary-db
- teriary-db comes up and syncs from secondary-db while keystone task
  runs on primary-db causing cluster to become desynced
- primary-db ends up trying to sync from the cluster and everything
  falls apart

The idea is to wait until the cluster has completely formed before
attempting to start openstack deployment. Additionally, this change
removes the one-by-one on the secondary database tasks which was added
as an attempt to address the cluster creation issues.

Change-Id: I4d62728922bb3a74576c419a934c1afd9187a35d
Closes-Bug: #1548271