Set default iptables INPUT policy to DROP
It was observed that on Fuel master node the default firewall INPUT
policy is set to 'ACCEPT' rather than to 'DROP'.
This leads to exposure of unnecessary services over a potentially
untrusted networks.

This patch updates default firewall INPUT policy to 'DROP'.

Patchset 3 fixes only indentation and styling issues.

Change-Id: I40f2cad53587a6025d0824390945962786e63cc7
Partial-Bug: #1524750